requirements.txt
Django>=3.0.8<3.1
djangorestframework>=3.11.0
drf-yasg>=1.17.1
social-auth-app-django>=3.4.0
boto3>=1.7.0
globus-sdk>=1.8.0
python-jose[cryptography]>=3.0.1
bdbag==1.3.0
minid>=2.0.0b3
